Undo compose package rename in datacapture
