+++ datacapture/src/androidTest/java/com/google/android/fhir/datacapture/test/QuestionnaireUiEspressoTest.kt
@@ -63,14 +63,14 @@ import com.google.android.fhir.datacapture.extensions.localDateTime
 import com.google.android.fhir.datacapture.validation.Invalid
 import com.google.android.fhir.datacapture.validation.QuestionnaireResponseValidator
 import com.google.android.fhir.datacapture.validation.Valid
-import com.google.android.fhir.datacapture.views.components.ADD_REPEATED_GROUP_BUTTON_TAG
-import com.google.android.fhir.datacapture.views.components.DATE_TEXT_INPUT_FIELD
-import com.google.android.fhir.datacapture.views.components.DELETE_REPEATED_GROUP_ITEM_BUTTON_TAG
-import com.google.android.fhir.datacapture.views.components.EDIT_TEXT_FIELD_TEST_TAG
-import com.google.android.fhir.datacapture.views.components.HANDLE_INPUT_DEBOUNCE_TIME
-import com.google.android.fhir.datacapture.views.components.PAGE_NAVIGATION_BUTTON_TAG
-import com.google.android.fhir.datacapture.views.components.REPEATED_GROUP_INSTANCE_HEADER_TITLE_TAG
-import com.google.android.fhir.datacapture.views.components.TIME_PICKER_INPUT_FIELD
+import com.google.android.fhir.datacapture.views.compose.ADD_REPEATED_GROUP_BUTTON_TAG
+import com.google.android.fhir.datacapture.views.compose.DATE_TEXT_INPUT_FIELD
+import com.google.android.fhir.datacapture.views.compose.DELETE_REPEATED_GROUP_ITEM_BUTTON_TAG
+import com.google.android.fhir.datacapture.views.compose.EDIT_TEXT_FIELD_TEST_TAG
+import com.google.android.fhir.datacapture.views.compose.HANDLE_INPUT_DEBOUNCE_TIME
+import com.google.android.fhir.datacapture.views.compose.PAGE_NAVIGATION_BUTTON_TAG
+import com.google.android.fhir.datacapture.views.compose.REPEATED_GROUP_INSTANCE_HEADER_TITLE_TAG
+import com.google.android.fhir.datacapture.views.compose.TIME_PICKER_INPUT_FIELD
 import com.google.android.fhir.datacapture.views.factories.NO_CHOICE_RADIO_BUTTON_TAG
 import com.google.android.fhir.datacapture.views.factories.YES_CHOICE_RADIO_BUTTON_TAG
 import com.google.android.material.progressindicator.LinearProgressIndicator
@@ -84,6 +84,7 @@ import kotlin.time.Duration.Companion.seconds
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.runBlocking
 import org.hamcrest.CoreMatchers
+import org.hamcrest.Matchers.allOf
 import org.hl7.fhir.r4.model.DateTimeType
 import org.hl7.fhir.r4.model.DateType
 import org.hl7.fhir.r4.model.Questionnaire
