+++ datacapture/src/androidTest/java/com/google/android/fhir/datacapture/test/views/DialogMultiSelectViewHolderFactoryTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023-2026 Google LLC
+ * Copyright 2023-2025 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -48,10 +48,10 @@ import com.google.android.fhir.datacapture.validation.Invalid
 import com.google.android.fhir.datacapture.validation.NotValidated
 import com.google.android.fhir.datacapture.views.QuestionTextConfiguration
 import com.google.android.fhir.datacapture.views.QuestionnaireViewItem
-import com.google.android.fhir.datacapture.views.components.ERROR_TEXT_AT_HEADER_TEST_TAG
-import com.google.android.fhir.datacapture.views.components.OPTION_CHOICE_LIST_TAG
-import com.google.android.fhir.datacapture.views.components.OPTION_CHOICE_TAG
-import com.google.android.fhir.datacapture.views.components.OTHER_OPTION_TEXT_FIELD_TAG
+import com.google.android.fhir.datacapture.views.compose.ERROR_TEXT_AT_HEADER_TEST_TAG
+import com.google.android.fhir.datacapture.views.compose.OPTION_CHOICE_LIST_TAG
+import com.google.android.fhir.datacapture.views.compose.OPTION_CHOICE_TAG
+import com.google.android.fhir.datacapture.views.compose.OTHER_OPTION_TEXT_FIELD_TAG
 import com.google.android.fhir.datacapture.views.factories.DialogSelectViewHolderFactory
 import com.google.android.fhir.datacapture.views.factories.MULTI_SELECT_TEXT_FIELD_TAG
 import com.google.android.fhir.datacapture.views.factories.QuestionnaireItemViewHolder
