+++ datacapture/src/androidTest/java/com/google/android/fhir/datacapture/test/views/GroupViewHolderFactoryTest.kt
@@ -31,8 +31,8 @@ import com.google.android.fhir.datacapture.test.TestActivity
 import com.google.android.fhir.datacapture.validation.Invalid
 import com.google.android.fhir.datacapture.validation.NotValidated
 import com.google.android.fhir.datacapture.views.QuestionnaireViewItem
-import com.google.android.fhir.datacapture.views.components.ERROR_TEXT_TAG
-import com.google.android.fhir.datacapture.views.components.HEADER_TAG
+import com.google.android.fhir.datacapture.views.compose.ERROR_TEXT_TAG
+import com.google.android.fhir.datacapture.views.compose.HEADER_TAG
 import com.google.android.fhir.datacapture.views.factories.GroupViewHolderFactory
 import com.google.android.fhir.datacapture.views.factories.QuestionnaireItemViewHolder
 import com.google.common.truth.Truth.assertThat
