+++ datacapture/src/main/java/com/google/android/fhir/datacapture/views/compose/ChoiceCheckbox.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025-2026 Google LLC
+ * Copyright 2025 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.datacapture.views.components
+package com.google.android.fhir.datacapture.views.compose
 
 import android.graphics.drawable.Drawable
 import androidx.compose.foundation.background
@@ -24,11 +24,12 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.selection.selectable
+import androidx.compose.foundation.selection.toggleable
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Checkbox
+import androidx.compose.material3.CheckboxDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.RadioButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -44,16 +45,16 @@ import androidx.core.graphics.drawable.toBitmap
 import com.google.android.fhir.datacapture.R
 
 @Composable
-internal fun ChoiceRadioButton(
+internal fun ChoiceCheckbox(
   label: AnnotatedString,
-  selected: Boolean,
+  checked: Boolean,
   enabled: Boolean,
   modifier: Modifier = Modifier,
   image: Drawable? = null,
-  onClick: () -> Unit,
+  onCheckedChange: (Boolean) -> Unit,
 ) {
   val backgroundColor =
-    if (selected) {
+    if (checked) {
       MaterialTheme.colorScheme.primary.copy(alpha = 0.12f)
     } else {
       MaterialTheme.colorScheme.surface
@@ -61,14 +62,14 @@ internal fun ChoiceRadioButton(
 
   val borderColor = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.15f)
   val textColor =
-    if (selected) {
+    if (checked) {
       MaterialTheme.colorScheme.primary
     } else {
       MaterialTheme.colorScheme.onSurface
     }
 
   val shape =
-    if (selected) {
+    if (checked) {
       RoundedCornerShape(4.dp)
     } else {
       RoundedCornerShape(8.dp)
@@ -80,17 +81,17 @@ internal fun ChoiceRadioButton(
         .clip(shape)
         .background(backgroundColor)
         .then(
-          if (!selected) {
+          if (!checked) {
             Modifier.border(1.dp, borderColor, shape)
           } else {
             Modifier
           },
         )
-        .selectable(
-          selected = selected,
+        .toggleable(
+          value = checked,
           enabled = enabled,
-          role = Role.RadioButton,
-          onClick = onClick,
+          role = Role.Checkbox,
+          onValueChange = onCheckedChange,
         )
         .padding(
           start = dimensionResource(R.dimen.option_item_between_text_and_icon_padding),
@@ -100,10 +101,16 @@ internal fun ChoiceRadioButton(
         ),
     verticalAlignment = Alignment.CenterVertically,
   ) {
-    RadioButton(
-      selected = selected,
-      onClick = null,
+    Checkbox(
+      checked = checked,
+      onCheckedChange = null,
       enabled = enabled,
+      colors =
+        CheckboxDefaults.colors(
+          checkedColor = MaterialTheme.colorScheme.primary,
+          uncheckedColor = MaterialTheme.colorScheme.onSurface,
+          checkmarkColor = MaterialTheme.colorScheme.surface,
+        ),
     )
     // Display image
     image?.let { drawable ->
@@ -111,7 +118,7 @@ internal fun ChoiceRadioButton(
       Icon(
         bitmap = drawable.toBitmap().asImageBitmap(),
         contentDescription = null,
-        modifier = Modifier.testTag(CHOICE_RADIO_BUTTON_IMAGE_TAG).size(24.dp),
+        modifier = Modifier.testTag(CHOICE_CHECKBOX_IMAGE_TAG).size(24.dp),
       )
     }
     Spacer(
@@ -126,4 +133,4 @@ internal fun ChoiceRadioButton(
   }
 }
 
-const val CHOICE_RADIO_BUTTON_IMAGE_TAG = "radio_button_option_icon"
+const val CHOICE_CHECKBOX_IMAGE_TAG = "checkbox_option_icon"
