+++ datacapture/src/main/java/com/google/android/fhir/datacapture/views/compose/OptionDialogSelect.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025-2026 Google LLC
+ * Copyright 2025 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.datacapture.views.components
+package com.google.android.fhir.datacapture.views.compose
 
 import android.content.Context
 import android.graphics.drawable.Drawable
