+++ datacapture/src/main/java/com/google/android/fhir/datacapture/views/compose/RepeatedGroupAddItem.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.fhir.datacapture.views.components
+package com.google.android.fhir.datacapture.views.compose
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
