+++ datacapture/src/main/java/com/google/android/fhir/datacapture/views/factories/EditTextViewHolderDelegate.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022-2026 Google LLC
+ * Copyright 2022-2025 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -35,11 +35,11 @@ import com.google.android.fhir.datacapture.extensions.itemMedia
 import com.google.android.fhir.datacapture.extensions.localizedFlyoverAnnotatedString
 import com.google.android.fhir.datacapture.extensions.unit
 import com.google.android.fhir.datacapture.views.QuestionnaireViewItem
-import com.google.android.fhir.datacapture.views.components.EditTextFieldItem
-import com.google.android.fhir.datacapture.views.components.EditTextFieldState
-import com.google.android.fhir.datacapture.views.components.Header
-import com.google.android.fhir.datacapture.views.components.MediaItem
-import com.google.android.fhir.datacapture.views.components.UnitText
+import com.google.android.fhir.datacapture.views.compose.EditTextFieldItem
+import com.google.android.fhir.datacapture.views.compose.EditTextFieldState
+import com.google.android.fhir.datacapture.views.compose.Header
+import com.google.android.fhir.datacapture.views.compose.MediaItem
+import com.google.android.fhir.datacapture.views.compose.UnitText
 import kotlinx.coroutines.Dispatchers
 
 class EditTextViewHolderDelegate(
